requirements.txt
openai>=1.3.5
langchain>=0.1.13
langchain-community>=0.0.24
python-dotenv
streamlit
